Still need to fix use of probid >= 1000000 for clarification
categories.

Conflicts resolved:
	lib/lib.misc.php
	lib/relations.php
	lib/www/clarification.php
	lib/www/print.php
	lib/www/scoreboard.php
	sql/mysql_db_structure.sql
	www/jury/submission.php
	www/team/clarification.php

+++ lib/www/common.jury.php
@@ -163,7 +163,7 @@ if (!function_exists('parse_ini_string')) {
  */
 function importZippedProblem($zip, $probid = NULL)
 {
-	global $DB;
+	global $DB, $teamid;
 	$prop_file = 'domjudge-problem.ini';
 
 	$ini_keys = array('probid', 'cid', 'name', 'allow_submit', 'allow_judge',
@@ -198,7 +198,7 @@ function importZippedProblem($zip, $probid = NULL)
 			unset($ini_array['probid']);
 			unset($ini_array['cid']);
 
-			$DB->q('UPDATE problem SET %S WHERE probid = %s', $ini_array, $probid);
+			$DB->q('UPDATE problem SET %S WHERE probid = %i', $ini_array, $probid);
 		}
 	}
 
@@ -206,7 +206,7 @@ function importZippedProblem($zip, $probid = NULL)
 	foreach (array('pdf', 'html', 'txt') as $type) {
 		$text = $zip->getFromName('problem.' . $type);
 		if ($text !== FALSE) {
-			$DB->q('UPDATE problem SET problemtext = %s, problemtext_type = %s WHERE probid = %s',
+			$DB->q('UPDATE problem SET problemtext = %s, problemtext_type = %s WHERE probid = %i',
 				$text, $type, $probid);
 			break;
 		}
@@ -214,7 +214,7 @@ function importZippedProblem($zip, $probid = NULL)
 
 	// Insert/update testcases
 	$maxrank = 1 + $DB->q('VALUE SELECT max(rank) FROM testcase
-	                       WHERE probid = %s', $probid);
+	                       WHERE probid = %i', $probid);
 	for ($j = 0; $j < $zip->numFiles; $j++) {
 		$filename = $zip->getNameIndex($j);
 		if ( ends_with($filename, ".in") ) {
@@ -226,7 +226,7 @@ function importZippedProblem($zip, $probid = NULL)
 
 				$DB->q('INSERT INTO testcase (probid, rank,
 				        md5sum_input, md5sum_output, input, output, description)
-				        VALUES (%s, %i, %s, %s, %s, %s, %s)',
+				        VALUES (%i, %i, %s, %s, %s, %s, %s)',
 				       $probid, $maxrank, md5($testin), md5($testout),
 				       $testin, $testout, $basename);
 				$maxrank++;
@@ -235,7 +235,7 @@ function importZippedProblem($zip, $probid = NULL)
 	}
 
 	// submit reference solutions
-	if ( $DB->q('VALUE SELECT allow_submit FROM problem WHERE probid = %s', $probid) ) {
+	if ( $DB->q('VALUE SELECT allow_submit FROM problem WHERE probid = %i', $probid) ) {
 		// First find all submittable languages:
 		$langs = $DB->q('KEYVALUETABLE SELECT langid, extensions
  		                 FROM language WHERE allow_submit = 1');
@@ -250,13 +250,13 @@ function importZippedProblem($zip, $probid = NULL)
 					break;
 				}
 			}
-			if( !empty($langid) ) {
+			if( !empty($langid) && !empty($teamid) ) {
 				if ( !($tmpfname = mkstemps(TMPDIR."/ref_solution-XXXXXX",0)) ) {
 					error("Could not create temporary file.");
 				}
 				file_put_contents($tmpfname, $zip->getFromIndex($j));
 				if( filesize($tmpfname) <= dbconfig_get('sourcesize_limit')*1024 ) {
-					submit_solution('domjudge', $probid, $langid, array($tmpfname), array($filename));
+					submit_solution($teamid, $probid, $langid, array($tmpfname), array($filename));
 				}
 				unlink($tmpfname);
 			}
